+++ package.json
@@ -26,6 +26,7 @@
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1",
+    "react-transition-group": "^4.4.5",
     "sass": "^1.63.6",
     "sass-loader": "^13.3.2",
     "typescript": "^4.9.5",
