+++ src/components/Loader/style.scss
@@ -1,0 +1,103 @@
+.spinner {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background: #fff;
+  font-size: 1.5rem;
+}
+
+@keyframes ldio-ogtm8crjy8g {
+  0% {
+    transform: translate(12px, 80px) scale(0);
+  }
+  25% {
+    transform: translate(12px, 80px) scale(0);
+  }
+  50% {
+    transform: translate(12px, 80px) scale(1);
+  }
+  75% {
+    transform: translate(80px, 80px) scale(1);
+  }
+  100% {
+    transform: translate(148px, 80px) scale(1);
+  }
+}
+@keyframes ldio-ogtm8crjy8g-r {
+  0% {
+    transform: translate(148px, 80px) scale(1);
+  }
+  100% {
+    transform: translate(148px, 80px) scale(0);
+  }
+}
+@keyframes ldio-ogtm8crjy8g-c {
+  0% {
+    background: #832bc1;
+  }
+  25% {
+    background: #1bd993;
+  }
+  50% {
+    background: #096cfe;
+  }
+  75% {
+    background: #503ae6;
+  }
+  100% {
+    background: #832bc1;
+  }
+}
+.ldio-ogtm8crjy8g div {
+  position: absolute;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  transform: translate(80px, 80px) scale(1);
+  background: #832bc1;
+  animation: ldio-ogtm8crjy8g 2.2222222222222223s infinite cubic-bezier(0, 0.5, 0.5, 1);
+}
+.ldio-ogtm8crjy8g div:nth-child(1) {
+  background: #503ae6;
+  transform: translate(148px, 80px) scale(1);
+  animation: ldio-ogtm8crjy8g-r 0.5555555555555556s infinite cubic-bezier(0, 0.5, 0.5, 1),
+    ldio-ogtm8crjy8g-c 2.2222222222222223s infinite step-start;
+}
+.ldio-ogtm8crjy8g div:nth-child(2) {
+  animation-delay: -0.5555555555555556s;
+  background: #832bc1;
+}
+.ldio-ogtm8crjy8g div:nth-child(3) {
+  animation-delay: -1.1111111111111112s;
+  background: #503ae6;
+}
+.ldio-ogtm8crjy8g div:nth-child(4) {
+  animation-delay: -1.6666666666666665s;
+  background: #096cfe;
+}
+.ldio-ogtm8crjy8g div:nth-child(5) {
+  animation-delay: -2.2222222222222223s;
+  background: #1bd993;
+}
+.loadingio-spinner-ellipsis-57ekl48z0o3 {
+  width: 200px;
+  height: 200px;
+  display: inline-block;
+  overflow: hidden;
+  background: #fff;
+}
+.ldio-ogtm8crjy8g {
+  width: 100%;
+  height: 100%;
+  position: relative;
+  transform: translate(0) scale(1);
+  backface-visibility: hidden;
+  transform-origin: 0 0; /* see note above */
+}
+.ldio-ogtm8crjy8g div {
+  box-sizing: content-box;
+}
