+++ src/components/SignUp/CheckConfirmCode.tsx
@@ -1,48 +1,18 @@
-import { Box, Button, Typography } from "@mui/material";
-import { Formik, Field, Form, ErrorMessage } from "formik";
+import { Box, Button } from "@mui/material";
+import { Formik, Form } from "formik";
 import * as Yup from "yup";
-import { Link, useNavigate } from "react-router-dom";
+import { Link, useNavigate, useLocation } from "react-router-dom";
 import { useState } from "react";
-import { useLocation } from "react-router-dom";
 import { IoIosArrowBack } from "react-icons/io";
+import { ThemeButtons } from "../helpers/theme-buttons";
+import { FormField } from "../helpers/form-field";
+import { FormSubmit } from "../helpers/form-submit";
+
 const validationSchema = Yup.object().shape({
-  code: Yup.number()
-    .min(3, "Code must be at least 3 characters")
-    // .max(7, "Max 7")
+  code: Yup.string()
+    .min(3, "Code must have at least 3 characters")
     .required("Code is required"),
 });
-const themes = [
-  {
-    background: "#1A1A2E",
-    color: "#FFFFFF",
-    primaryColor: "#0F3460",
-  },
-  {
-    background: "#461220",
-    color: "#FFFFFF",
-    primaryColor: "#E94560",
-  },
-  {
-    background: "#192A51",
-    color: "#FFFFFF",
-    primaryColor: "#967AA1",
-  },
-  {
-    background: "#F7B267",
-    color: "#000000",
-    primaryColor: "#F4845F",
-  },
-  {
-    background: "#F25F5C",
-    color: "#000000",
-    primaryColor: "#642B36",
-  },
-  {
-    background: "#231F20",
-    color: "#FFF",
-    primaryColor: "#BB4430",
-  },
-];
 
 export const CheckConfirmCode = () => {
   const location = useLocation();
@@ -50,7 +20,7 @@ export const CheckConfirmCode = () => {
   const email = searchParams.get("email");
   const navigate = useNavigate();
   const [showError, setShowError] = useState("");
-  const [selectedTheme, setSelectedTheme] = useState("");
+  const [selectedTheme, setSelectedTheme] = useState("#231f20");
 
   const checkConfirmationCode = async (values: any) => {
     try {
@@ -61,6 +31,7 @@ export const CheckConfirmCode = () => {
           body: JSON.stringify({ email, code: values.code }),
         }
       );
+
       if (response.ok) {
         const data = response.json();
         navigate(`/set-password?email=${email}`);
@@ -70,11 +41,12 @@ export const CheckConfirmCode = () => {
         const leftTriesCount = errorData?.left_tries_count;
         const textError = `${errorMessage} ${leftTriesCount}`;
         setShowError(textError);
+        
       }
 
-      console.log("Confirmation code check successful");
+      console.log("Проверка кода подтверждения выполнена успешно");
     } catch (error) {
-      console.error("Error checking confirmation code", error);
+      console.error("Ошибка при проверке кода подтверждения", error);
     }
   };
 
@@ -101,63 +73,30 @@ export const CheckConfirmCode = () => {
               alt="illustration"
               className="illustration"
             />
-            <h1 className="opacity form-title">Сonfirm mail</h1>
+            <h1 className="opacity form-title">Confirm mail</h1>
             <Formik
               initialValues={{ code: "" }}
               validationSchema={validationSchema}
               onSubmit={checkConfirmationCode}
             >
               <Form>
-                <Field
-                  type="text"
-                  name="username"
+                <FormField
+                  name="usernamecode"
                   placeholder="Email"
                   readOnly
                   value={email}
                 />
-                <ErrorMessage
-                  name="username"
-                  component="p"
-                  className="form-error"
-                />
-                <Field type="text" name="code" placeholder="Code" />
-                <ErrorMessage
-                  name="code"
-                  component="p"
-                  className="form-error"
-                />
-                <Button type="submit" className="opacity">
-                  Next
-                </Button>
+                <FormField name="code" placeholder="Code" />
+                <FormSubmit buttonText="Next" />
               </Form>
             </Formik>
-            {/* <div className="register-forget opacity">
-              <a href="">REGISTER</a>
-              <a href="">FORGOT PASSWORD</a>
-            </div> */}
           </div>
           <div className="circle circle-two"></div>
         </div>
-        <div className="theme-btn-container">
-          {themes.map((theme: any) => {
-            const style = {
-              background: theme.background,
-              marginBottom: "5px",
-              cursor: "pointer",
-              padding: "10px",
-            };
-            return (
-              <span
-                style={style}
-                key={theme.background}
-                onClick={() => handleThemeChange(theme.background)}
-              >
-                {theme.background}
-              </span>
-            );
-          })}
-        </div>
+        <ThemeButtons handleThemeChange={handleThemeChange} />
       </Box>
     </Box>
   );
 };
+
+export default CheckConfirmCode;
