+++ src/components/helpers/form-submit.tsx
@@ -1,0 +1,9 @@
+// FormSubmit.js
+import { Button } from "@mui/material";
+
+export const FormSubmit = ({ buttonText }:any) => (
+  <Button type="submit" className="opacity">
+    {buttonText}
+  </Button>
+);
+
