+++ src/index.tsx
@@ -3,13 +3,17 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
+import { Suspense } from "react";
+import Loader from "./components/Loader/Loader";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <App />
+    <Suspense fallback={<Loader text="Loading ..." />}>
+      <App />
+    </Suspense>
   </React.StrictMode>
 );
